ZON-4064: Move write to brightcove handlers to the update module, so the symmetry is clearer
